requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.91.0
h11==0.14.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
spotipy==2.22.1
starlette==0.24.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
